- Dragging and Clicking are functional in the GUI.
A few things to note: we no longer seem to have the ability to determine if events are within the control. The work around for this is using isPointInRect() for each Control's onEvent() functions -including- its derivative controls.  We will need to think of a better mechanism to work with the new Event system.